fix: 디렉토리 구조, oracle datasource

+++ src/main/java/com/yorizip/myapp/user/service/UserService.java
@@ -26,7 +26,7 @@ public class UserService {
     private static final Logger logger = LoggerFactory.getLogger(UserService.class);
 
     @Transactional
-    public int saveUser(UserVO user) {
+    public Long saveUser(UserVO user) {
         userMapper.insertUser(user);
         return user.getUserId();
     }
@@ -242,7 +242,7 @@ public class UserService {
 
         // User 정보 저장
         userMapper.insertUser(user);
-        int userId = user.getUserId(); // DB에 저장된 후, 생성된 userId를 가져옴
+        Long userId = user.getUserId(); // DB에 저장된 후, 생성된 userId를 가져옴
         logger.info("Received 4: {}", userEmail);
 
         // 소셜 로그인 정보 처리
@@ -278,7 +278,7 @@ public class UserService {
 
         // User 정보 저장
         userMapper.insertUser(user);
-        int userId = user.getUserId(); // DB에 저장된 후, 생성된 userId를 가져옴
+        Long userId = user.getUserId(); // DB에 저장된 후, 생성된 userId를 가져옴
         logger.info("Received 4: {}", userEmail);
 
         return user;
